% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{buildPopsURL}
\alias{buildPopsURL}
\title{Set new url to find Populations}
\usage{
buildPopsURL(character)
}
\arguments{
\item{character}{the path to the populations via API}
}
\value{
This will return new url for finding populations
}
\description{
Set new url to find Populations
}
\examples{
\dontrun{
url <- "http://localhost:4567/api/v1/populations/all?workspaceid="
urlPops <- buildPopsURL(url)
} 
}
