% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{buildSamURL}
\alias{buildSamURL}
\title{Set new url based on loaded workspace to find Samples loaded.}
\usage{
buildSamURL(character)
}
\arguments{
\item{character}{This will become the new url to find samples based on loaded workspace.}
}
\value{
This will be new url to find samples.
}
\description{
Set new url based on loaded workspace to find Samples loaded.
}
\examples{
\dontrun{
getWorkspaces(url)
urlSams <- buildSamURL("http://localhost:4567/api/v1/samples/all?workspaceid=")
} 
}
