% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableOjectiveTypes.R
\name{getAvailableObjectiveTypes}
\alias{getAvailableObjectiveTypes}
\title{Returns possible objective types.}
\usage{
getAvailableObjectiveTypes()
}
\value{
[\code{character}] Vector of possible objective types.
}
\description{
Returns possible objective types.
}

