% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.getters.R
\name{getNoirID}
\alias{getNoirID}
\title{Get ID of an noir optimizer, e.g., short name of an optimizer or the optimizer
which produces a result object.}
\usage{
getNoirID(x)
}
\arguments{
\item{x}{[object]\cr
An optimizer or a result object.}
}
\value{
[\code{character(1)}]
}
\description{
Get ID of an noir optimizer, e.g., short name of an optimizer or the optimizer
which produces a result object.
}

