% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{getParameters}
\alias{getParameters}
\title{Make call to API for Parameters Associated with Samples loaded into a workspace}
\usage{
getParameters(character)
}
\arguments{
\item{character}{The path to parameters via API}
}
\description{
Make call to API for Parameters Associated with Samples loaded into a workspace
}
\examples{
\dontrun getParameters(urlParams)
}
