% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{getPopulations}
\alias{getPopulations}
\title{Make a call to API to get Populations associated with Samples}
\usage{
getPopulations(character)
}
\arguments{
\item{character}{the url path for finding populations in workspace}
}
\value{
This will return Populations from a workspace
}
\description{
Make a call to API to get Populations associated with Samples
}
\examples{
\dontrun{
urlPops <- buildPopsURL(url)
getPopulations(urlPops)
} 
}
