% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{getSamples}
\alias{getSamples}
\title{Make call to API for Samples Names loaded into a workspace}
\usage{
getSamples(character)
}
\arguments{
\item{character}{The path to workspace via API}
}
\description{
Make call to API for Samples Names loaded into a workspace
}
\examples{
\dontrun{
urlSams <- buildSamURL("http://localhost:4567/api/v1/samples/all?workspaceid=")
getSamples(urlSams)
}
}
