% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedTags.R
\name{getSupportedTags}
\alias{getSupportedTags}
\title{Returns currently valid tags.}
\usage{
getSupportedTags()
}
\value{
[\code{character}] Vector of possible tags.
}
\description{
Returns currently valid tags.
}

