% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{getWorkspaces}
\alias{getWorkspaces}
\title{Noir}
\source{
This code could be made faster by not writing out the data frames and
perhaps by streaming directly from the API.
}
\usage{
getWorkspaces(character)
}
\arguments{
\item{character}{The url path to workspace via API}
}
\value{
The Noir code will make call to API to gather and then create data frame of connected
workspace names, samples, parameters, and populations. Otherwise returns error message.
}
\description{
Make a call to API to find loaded workspaces.
}
\note{
This code can be used to feed other plugins that would require this info.
}
\examples{
\dontrun dfWorkspaces <- getWorkspaces(url)
}
