% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasProperties.R
\name{hasProperties}
\alias{hasProperties}
\title{Check if optimizer has certain properties.}
\usage{
hasProperties(optimizer, properties)
}
\arguments{
\item{optimizer}{[\code{noir_optimizer}]\cr
The optimizer.}

\item{properties}{[\code{character}]\cr
Properties/tags to check for.
If multiple properties are are passed it is checked whether the optimizer
has all of them.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if optimizer has certain properties.
}

