% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internallyRegisterOptimizer.R
\name{internallyRegisterOptimizer}
\alias{internallyRegisterOptimizer}
\title{Register a new optimizer / optimization algorithm.}
\usage{
internallyRegisterOptimizer(id, class, packages, hyper.par.set,
  par.values = list(), objective.type = "single-objective", tags = NULL)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
String representation of the optimizer.}

\item{class}{[\code{character(1)}]\cr
Class used internally for the optimizer. Basically this can be equal to \code{id}.}

\item{packages}{[\code{character}*]\cr
Packages needed to execute the optimizer.
Default is \code{NULL}, i.e., base R is sufficient.}

\item{hyper.par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set defining the algorithm parameters.
See \code{\link[ParamHelpers]{makeParamSet}} for details.}

\item{par.values}{[\code{list}]\cr
Named list of parameter values.
Default is the empty list.}

\item{objective.type}{[\code{character(1)}]\cr
One of \dQuote{single-objective} or \dQuote{multi-objective}.
See \code{\link{getAvailableObjectiveTypes}}.}

\item{tags}{[\code{character(1)}]\cr
Keywords which characterize the properties of the optimizer.
See \code{\link{getSupportedTags}} for a list of available tags.
Default is \code{NULL}, i.e., no tags are assigned.}
}
\value{
[\code{optimizer}] S3 object encapsulating the optimizer.
}
\description{
This function is the \dQuote{make} of \pkg{noir}. It is a constructor for
\pkg{noir}-optimizers.
}

