% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOptimizerParams.R
\name{setOptimizerParams}
\alias{setOptimizerParams}
\title{Set parameters of an optimizer.}
\usage{
setOptimizerParams(optimizer, ...)
}
\arguments{
\item{optimizer}{[\code{noir_optimizer}]\cr
Optimizer object.}

\item{...}{[any]\cr
Parameters to set.}
}
\value{
[\code{noir_optimizer}] with modified parameters.
}
\description{
Set parameters of an optimizer.
}
\examples{
optimizer = setUpOptimizer("neldermead")
optimizer = setOptimizerParams(optimizer, alpha = 3.1, maxit = 1000)
}

