% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpOptimizer.R
\name{setUpOptimizer}
\alias{setUpOptimizer}
\title{Generate an optimizer object.}
\usage{
setUpOptimizer(class, ...)
}
\arguments{
\item{class}{[\code{character(1)}]\cr
ID, i.e., short name, of the optimizer to create.}

\item{...}{[any]\cr
Further parameters for the optimizer.
If none provided, the default will be used. Later on, one can set the parameters
in the \code{\link{noir}} function.}
}
\value{
[\code{noir_optimizer}]
}
\description{
Creates an object of type \code{noir_optimizer}.
}
\seealso{
\code{\link{noir}}
}

