% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProperties.R
\name{getProperties}
\alias{getProperties}
\title{Return properties/tags of an optimizer.}
\usage{
getProperties(optimizer)
}
\arguments{
\item{optimizer}{[\code{noir_optimizer}]\cr
The optimizer.}
}
\value{
[\code{character}] Character vector of properties.
}
\description{
Return properties/tags of an optimizer.
}

