% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noir.R
\name{getWorkspaces}
\alias{getWorkspaces}
\title{Noir

Make a call to API to find loaded workspaces.}
\usage{
getWorkspaces(character)
}
\arguments{
\item{character}{The url path to workspace via API}
}
\value{
The Noir code will make call to API to gather and then create data frame of connected
workspace names, samples, parameters, and populations. Otherwise returns error message.
}
\description{
Noir

Make a call to API to find loaded workspaces.
}
\note{
This code can be used to feed other plugins that would require this info.
This code could be modified to stream info directly from the API and not writing  
out the data frames.
}
\examples{
\dontrun{
urlWSP <- "http://localhost:4567/api/v1/workspaces/all"
getWorkspaces(urlWSP)
} 
}
