% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noir.R
\name{noir}
\alias{noir}
\title{Run an optimizer on an objective function.}
\usage{
noir(optimizer, obj.fn, ...)
}
\arguments{
\item{optimizer}{[\code{noir_optimizer}]\cr
Noir optimizer.}

\item{obj.fn}{[\code{smoof_function}]\cr
Objective function.}

\item{...}{[any]\cr}
}
\value{
[\code{noir_result}]
}
\description{
Expects an \code{noir_optimizer} and a \code{smoof_function} and applies the
optimizer to the function.
}
\examples{
optimizer = setUpOptimizer("neldermead")
res = noir(optimizer, smoof::makeSphereFunction(2L), alpha = 1.5)
}

