% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpOptimizer.R
\name{setUpOptimizer}
\alias{setUpOptimizer}
\title{Generate an optimizer object.}
\usage{
setUpOptimizer(class)
}
\arguments{
\item{class}{[\code{character(1)}]\cr
ID, i.e., short name, of the optimizer to create.}
}
\value{
[\code{noir_optimizer}]
}
\description{
Creates an object of type \code{noir_optimizer}.
}
\seealso{
\code{\link{noir}}
}

